feat: add Anime class and comparators for sorting by title and episodes

+++ maratona-java/src/javacore/Npolimorfismo/dominio/Computador.java
@@ -4,7 +4,7 @@ public class Computador extends Produto{
     public static final double IMPOSTO_POR_CENTO =0.21;
 
     public Computador(String nome, double valor) {
-        super(nome, valor);
+        super(nome, (int)valor);
     }
 
     @Override
