+++ maratona-java/src/javacore/Npolimorfismo/dominio/Produto.java
@@ -3,9 +3,9 @@ package javacore.Npolimorfismo.dominio;
 public abstract class Produto implements Taxavel {
     // abstract aquele que "nao" pode existir e criar instancias diretadmente dela
     protected String nome;
-    protected double valor;
+    protected int valor;
 
-    public Produto(String nome, double valor) {
+    public Produto(String nome, int valor) {
         this.nome = nome;
         this.valor = valor;
     }
