+++ maratona-java/src/javacore/Npolimorfismo/dominio/Televisao.java
@@ -3,7 +3,7 @@ package javacore.Npolimorfismo.dominio;
 public class Televisao extends Produto {
     public static final double IMPOSTO_POR_CENTO =0.21;
     public Televisao(String nome, double valor) {
-        super(nome, valor);
+        super(nome, (int)valor);
     }
 
     @Override
